+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+### Changed
+ - Close button now closes settings if settings are up.
+
 ## [1.1.0] - 2021-11-20
 ### Removed
  - Default title bar, replaced with custom one.